server/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytest-django==4.1.0
channels==3.0.2
channels-redis==3.2.0
pytest-asyncio==0.14.0